import Qt.labs.templates 1.0 => QtQuick.Templates 2.0
Even though there never was such thing is QtQuick.Templates 1.0,
we'll use version 2.0 to keep it in sync with QtQuick.Controls 2.0
(next step). The templates define (and in the future, revision) the
API of the controls, so the two must have the same version or it
will get messy.

Docs and plugins.qmltypes will be updated in follow up commits.

Change-Id: Idc2e76f32f348ba35eb4b1fa167c6868300aa910
Task-number: QTBUG-52549
Reviewed-by: J-P Nurmi <jpnurmi@qt.io>