Introduce a 'System' theme to the styles.
Setting the theme to System chooses either the light or dark theme
based on the system theme colors. However, when reading the value of
the theme property, the value is never System, but the actual theme.

[ChangeLog][Controls][Material] Added Material.System theme enum value,
that can be used to let the Material style choose either the light or
dark theme based on the system theme colors.
[ChangeLog][Controls][Universal] Added Universal.System theme enum value,
that can be used to let the Universal style choose either the light or
dark theme based on the system theme colors.

Change-Id: Ibfc9f01953cb8322b64d59413cfbaef9d4bb28fd
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>