QQuickWindowPrivate: Only update transform on polish if needed
As it stood, whenever QtQuick did a polish we would update
IM transform. The result would be that we signaled changes
to e.g cursor rectangle and input rectangle each time the
cursor was redrawn. This of course caused code elsewhere, e.g
in the platform plugins, to recalculate overlays such as
cursor handles when nothing with regards to input item
transform had actually changed.

This patch will add some extra checks that the effective transform
of the focus object has really changed before telling IM
to update.

Change-Id: If7057e4dd8f41e251a27d68fcaebdb10da953ee7
Reviewed-by: Jan Arve Sæther <jan-arve.saether@theqtcompany.com>