Introduce RendererInfo
Modeled after OpenGLInfo, RendererInfo.api is an easy to use attached
property that reports the current graphics API as long as the item is
associated with a window that has the scenegraph initialized.

Applications should not worry about this in general since they should
target one given API, however tests, examples and some advanced use cases
will find this handy. RendererInfo is also a great place to expose other
properties in the future.

Change-Id: Ie3411b8820fc4b74efd1b1d0143e936886f62d74
Reviewed-by: Andy Nichols <andy.nichols@qt.io>