Popup: use margins to determine whether to push inside window bounds
Previously, we were planning to introduce complex policy flags to make it
possible to specify whether a Popup is allowed to go outside the screen or not.
Using margins for that purpose avoids the need of introducing such dedicated
API just for that. From now on, margins are used to determine whether a popup
is pushed to fit inside the window bounds (margins removed), and implicit size
is used to determined whether a popup is resized to fit inside the window. In
other words, a popup with negative margins and no implicit size is allowed to
move outside the window bounds.

Change-Id: Ife1665755e4dae434751977ae3289eb1aa4f3c6d
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>