Material: fix SwitchDelegate press color
Elevation is required to get a shadow for the indicator, but it also
affects the press color, which is undesired. To avoid a darker press
color than the other delegates have, elevate only the indicator
instead of the whole delegate. Unfortunately, this results to two
Material attached object instances for the delegate, so it must be
blacklisted in tst_sanity...

Change-Id: I8b607ba8f7493fd34a4bdc131a5bfdc5a39159ae
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>