QML: Fix anchors on Windows.
When using an enum type for a bitfield (i.e. Anchors foo : 7), MSVC will
somehow store (or interpret) it as signed, while clang/gcc will do it
unsigned. This behavior can be made less exciting by specifying the
storage type of the enum.

However, as the exception that confirms the rule: gcc 4.8 on OpenSUSE 13
will complain that it can't store all Anchor values in a 7 bit bitfield,
but ONLY when a base type for an enum is specified.

Change-Id: I7514dd613017d321de55560affb9b355fa75fa2e
Reviewed-by: Simon Hausmann <simon.hausmann@theqtcompany.com>