Remove unncessary TODO in text material
The device pixel ratio is easily available from the engine always.
(how correct the value is, is another question, but it's not
the renderer's problem)

Differentiating between onscreen/offscreen is not needed here. The
original GL code is convoluted due to having to spport contexts with a
pbuffer surface (i.e. no real QWindow).

Change-Id: I5e5d42857ca00a53794ff3e3050bb256912e5bc6
Reviewed-by: Andy Nichols <andy.nichols@qt.io>