Popup: restore original size on reposition if appropriate
When popup need to reposition (for example, on screen orientation
change), it's resized to fit into window bounds. But if the implicit
size of it's content changed to accommodate window geometry, popup
wouldn't change it's size back.

This commit allow to change popup size when implicit width of the
content changed and new size could fit into window bounds.

Task-number: QTBUG-52729
Change-Id: I5a265b55e24495be9c1f2050cdcceddcdeb57974
Reviewed-by: J-P Nurmi <jpnurmi@qt.io>