Respect the modal property for Drawer
Even if Drawers default to modal true, is still possible to set them to
non modal at any time.
When the modal property is set to false in a Drawer instance, the
dimming gets disabled as expected, but it continues to steal clicks for
the whole area of the application window. This makes the handling of
the click on the overlay to check for the modal property before deciding
to steal the click.

Change-Id: I082453a0cc6dd5c1a3e0a123abc4ff018d56c789
Reviewed-by: J-P Nurmi <jpnurmi@qt.io>