Remove dependency of QQmlDesignerMetaObject to QQmlVMEMetaData
The constructor will use "fake" VME meta data if the provided QObject has a VME
meta object - that is the behavior of the vMEMetaDataForObject(object) call.
The QQmlVMEMetaObject constructor will set hasVMEMetaObject to true
unconditionally. By detecting the use of the fake meta data we can restore the
hasVMEMetaObject flag to what it was before. As I would like to remove the VME
meta-data, this patch removes the dependency to the meta data as a way of
detecting the situation where we have to restore the hasVMEMetaObject flag.
Instead this is done explicitly in the factor method that calls the constructor.

Change-Id: I9c5cbfbd8fb41ca65c597669ee04c1c5694dee2a
Reviewed-by: Lars Knoll <lars.knoll@theqtcompany.com>