Add an API to set the scenegraph backend
As an alternative to QT_QUICK_BACKEND and QMLSCENE_DEVICE.

The need for this is twofold: it is requested by Creator since environment
variables have potential issues with child processes inheriting them.
Also, platforms like WinRT may have issues with environment variables - as
such a concept may not necessarily exist.

Change-Id: Ia66b78dc83fce5af0f36d4b34dc094582eb6e61d
Reviewed-by: Robin Burchell <robin.burchell@viroteck.net>