Add QQuickWheelEvent::inverted()
When handling wheel events, it's useful to know whether or not
"natural" scrolling was enabled on platforms like OSX where
the scrolling direction can be reversed. It's especially useful
when implementing components like Slider (see QTBUG-35972).
This will soon be needed in QtQuick.Controls because we plan
to remove the WheelArea.

Change-Id: I2770e0aa39e7e54511d7f9f7cc12760580a55e5c
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>