Simplify id and alias collection
Ids and aliases that refer to their target by id are bound to the scope of
components. Instead of determining this boundary through a binary search for
bindings that refer to the inner object, we can simply query the flags of the
component object itself as to when to stop. This is based on the assumption
that a component has only one binding anyway.

Change-Id: I9a555ca95f8274b6d9a98793a2c946f47348f9d4
Reviewed-by: Lars Knoll <lars.knoll@theqtcompany.com>