D3D12: Remove m_lastNN members from text material
To reduce the data size. The data is anyway available in the element-dedicated
area of the constant buffer (better said, the plain CPU memory the
renderer uses to prepare the data for the engine), so compare with that
instead. Note that while the memcmp-memcpy pair may seem to have little
value at first, setting the UpdatedConstantBuffer flag only when needed does
have its benefits since the GPU-visible buffer mapping can be skipped completely
when nothing has changed.

Change-Id: Ie5d6797a06cb4259bd65114c5b69b11ff165e871
Reviewed-by: Andy Nichols <andy.nichols@qt.io>