Clean up property cache reference counting in QQmlVMEMetaObject
When the intercepting functionality of the VME meta-object was split into the
QQmlInterceptorMetaObject the property cache was moved as a member but the
manual refcounting remained in the QQmlVMEMetaObject constructor and
destructor. That's fragile and potential future sub-classes of the interceptor
would have to do the same refcounting. Instead let's just store a QQmlRefPointer
in the right place.

Similar logic applies to the QQmlDesignerMetaObject, which had a copy of the
property cache for now aparent reason.

Change-Id: I7f6d3917fc5ef1c4beb065525be10bb9c0fadcf3
Reviewed-by: Thomas Hartmann <Thomas.Hartmann@theqtcompany.com>
Reviewed-by: Lars Knoll <lars.knoll@qt.io>