Software Adaptation: Obey QQuickWindow::TextureIsOpaque flag
Force a image format conversion when requesting an Opaque texture.
There seems to be no way to do this with QPixmap::fromImage(...) so a
copy of the QImage must be made first which is unfortunate.  This is
further evidence that the Texture should be QImage based instead of
QPixmap based by default.

Change-Id: I89fb380e8e786d4ff93f01eb6466e118c2851f7a
Reviewed-by: Laszlo Agocs <laszlo.agocs@theqtcompany.com>