Universal: remove qdoc markers
The original plan was to include the default implementation (of all
styles, if we could make the snippets collapsiple) in the customization
docs. Since the default implementations tend to include internal types,
in the end we went the other way and wrote dedicated example snippets
for the customization docs. Therefore these markers no longer server a
purpose, but just add unnecessary extra bytes for the QML parser to
read.

Change-Id: I5740192b0a3e80614a795d32dabeb753f86bbd68
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>