Material: remove qdoc markers
The original plan was to include the default implementation (of all
styles, if we could make the snippets collapsiple) in the customization
docs. Since the default implementations tend to include internal types,
in the end we went the other way and wrote dedicated example snippets
for the customization docs. Therefore these markers no longer server a
purpose, but just add unnecessary extra bytes for the QML parser to
read.

Change-Id: Ic511d7ce7a032429626ee75517d69d21b21412a3
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>