Support glyphs outside em square in distance field cache
In the distance field cache, we would assume a max height of glyphs
at 64 pixels when the pixel size was 54 (em square size is 54 and
then some margin). This is an arbitrary assumption which does not
necessarily hold true. Indeed, some fonts would be clipped at
the bottom when assuming this.

[ChangeLog][QtQuick][Text] Fixed clipping of glyphs that extend
beyond font's em square.

Task-number: QTBUG-52389
Change-Id: I5f6c9be235d38841e40d1bd60fb380e4712dd41a
Reviewed-by: Yoann Lopes <yoann.lopes@qt.io>