fix incorrect propagation of hover events in some cases
Both the item itself and its descendants need to be considered together
when deciding whether or not it is interested in hover events.

A correction to f728a514abdf59e2ac556eef027a69d9f21db4b2.
Rename it while we're at it, to make the new meaning clearer.

Task-number: QTBUG-54019
Change-Id: I1fe9f4a7f71d037a542ffefd0f3aff6f52a80a40
Reviewed-by: Robin Burchell <robin.burchell@viroteck.net>