Gallery: fix the default style setting
A problem noticed in release testing:
Run the Gallery example without explicitly specifying any style. The
Gallery runs with the Default style, but the settings dialog wrongly
claims that it's running with the Universal style. The situation is
easy to emulate by removing ~/.config/QtProject/Gallery.conf (to
clear the previously stored style setting).

=> Fix the wrong default value (it was wrong because back in the days
the Universal style was used by default).

Change-Id: I7eb7565cb78e68cf5e85527e53724af217973584
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
Reviewed-by: Liang Qi <liang.qi@qt.io>