Add a tab widget to QQuickWidget example
...and include a tab with custom OpenGL rendering. Tab widgets are
interesting because they are commonly used in applications in combination
with widgets like QOpenGLWidget and QQuickWidget, and because they typically
trigger rapid show-hide sequences which QQuickWidget does not always handle
as well as it could.

Thus this serves both as a helpful example and a useful testing tool.

Plus it exposes at least two bugs already, namely that the continuous
animation in the QQuickFBO does not start when switching to the tab
containing the QQuickWidget in question, and that it is impossible
to integrate third-party rendering code in a robust manner due to
QQuickWidget destroying everything whenever hiding the widget. This
is not what happens with a QQuickView and the normal render loops.

Task-number: QTBUG-54133
Change-Id: Ie7d92cd0f685e4a26f4bb351cc023eb697a36bf4
Reviewed-by: Andy Nichols <andy.nichols@qt.io>