Don't be too eager to steal presses outside the drawer
While this is not a proper fix to the linked bug, debugging the issue revealed
that Drawer was way too eager to grab touch outside the drawer. It would not
do anything with the events until finger reached the drawer panel, so it's
completely unnecessary to steal the press far outside.

Change-Id: I76ff778b939327632cb29203db1c24e05c2a1a75
Task-number: QTBUG-53929
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>