ColorDialog & FontDialog: use widgets as a fallback
This makes these dialogs available on Windows and KDE. One more reason
to keep the module experimental. In the long run, platforms should
implement QPA dialogs so that we can remove the undesired dependency.

Also simple instance creation tests are included to make sure it
won't crash with non-QApplication apps.

Change-Id: Ib4dabe7b656a1aa1f114052a3f57c9b1232ff27f
Reviewed-by: J-P Nurmi <jpnurmi@qt.io>