Tumbler: change default visibleItemCount to 5
This looks much better than 3; it looks more like an actual Tumbler,
partly because the displacement shading is made more visible. This
should make it less confusing to interact with for first-time users.

In addition, we were previously setting it to 5 in several places.

[ChangeLog][Important Behavior Changes][Tumbler] Changed the default
value of visibleItemCount to 5 to make it visually clearer that it's a
Tumbler.

Change-Id: I7c16e05aa73f58e3405caa1826453cb824efae11
Reviewed-by: J-P Nurmi <jpnurmi@qt.io>